requirements.txt
stripe>=11.0.0
python-dotenv>=1.0.1